website/requirements.txt
fastapi[all]==0.115.0
uvicorn==0.31.0
jinja2==3.1.4
motor==3.6.0
redis==5.1.1
cbor2==5.6.4
google-cloud-vision==3.7.4
google-auth==2.35.0